Changed floats to int and numerics for consistency sake and attempt to fix not getting to confirmation screen on checkout.

+++ KevinBhavinRestaurant/ViewModel/ShoppingCartRemoveViewModel.cs
@@ -9,7 +9,7 @@ namespace KevinBhavinRestaurant.ViewModel
     public class ShoppingCartRemoveViewModel
     {
         public string Message { get; set; }
-        public float CartTotal { get; set; }
+        public decimal CartTotal { get; set; }
         public int CartCount { get; set; }
         public int ItemCount { get; set; }
         public int DeleteId { get; set; }
